% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkxandobj}
\alias{checkxandobj}
\title{check format of x and obj}
\usage{
checkxandobj(x, obj)
}
\arguments{
\item{x}{Data matrix input to other functions}

\item{obj}{Output of initializeISD}
}
\value{
Errors if needed
}
\description{
check format of x and obj
}
