% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterPerturbations.R
\name{clusterPerturbations}
\alias{clusterPerturbations}
\title{Cluster cells by their environment perturbations}
\usage{
clusterPerturbations(
  x,
  obj,
  cells = NULL,
  genes = "highlyperturbed",
  nclust = 12,
  residtype = "log2ratio",
  eps = 1
)
}
\arguments{
\item{x}{Expression matrix for complete dataset, same as was input to initializeISD. Cells in rows, genes in columns.}

\item{obj}{Output of initializeISD}

\item{cells}{Which cells to produce output for. If NULL, will return results for all cells.}

\item{genes}{Which genes to produce output for.
Default of "highlyperturbed" subsets to the most perturbed genes. If NULL, all genes will be used.}

\item{nclust}{How many clusters to fit}

\item{residtype}{Either "log2ratio" or "diff"}

\item{eps}{For log2ratio calculations, mean neighborhood expression will be thresholded below at this value}
}
\description{
Finds clusters of neighborhoods with similar perturbation patterns.
}
