% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{colorRightTailed}
\alias{colorRightTailed}
\title{Choose colors based on a vector of non-negative, right-tailed values}
\usage{
colorRightTailed(x, maxquant = 0.995, palette = c("grey80", "darkblue"))
}
\arguments{
\item{x}{Vector of non-negative values}

\item{maxquant}{x will be thresholded above at this quantile}

\item{palette}{Vector of colors, which will be fed into colorRampPalette}
}
\description{
Colors from 0-high, truncating values at a high quantile of the data so more
than just the most extreme outliers are visible. Use colorZeroCentered for vectors centered around 0.
}
