% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{colorZeroCentered}
\alias{colorZeroCentered}
\title{Choose colors based on a vector of non-negative, right-tailed values}
\usage{
colorZeroCentered(
  x,
  maxquant = 0.995,
  palette = c("darkblue", "blue", "grey80", "red", "darkred")
)
}
\arguments{
\item{x}{Vector of non-negative values}

\item{maxquant}{x will be thresholded above at this quantile}

\item{palette}{Vector of colors, which will be fed into colorRampPalette}
}
\description{
Colors from a vector centered around a meaningful 0, truncating values at high/low quantiles of the data so more
than just the most extreme outliers are visible
}
