% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatCellsArg}
\alias{formatCellsArg}
\title{Format "cells" argument to take the form of names}
\usage{
formatCellsArg(cells, x)
}
\arguments{
\item{cells}{Cells argument input to a larger function}

\item{x}{Data matrix input into a larger function}
}
\value{
Cells formatted as a vector of cell IDs
}
\description{
Format "cells" argument to take the form of names
}
