% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatGenesArg}
\alias{formatGenesArg}
\title{Format "genes" argument to take the form of names}
\usage{
formatGenesArg(genes, x)
}
\arguments{
\item{genes}{genes argument input to a larger function}

\item{x}{Data matrix input into a larger function}
}
\value{
Cells formatted as a vector of cell IDs
}
\description{
Format "genes" argument to take the form of names
}
