% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGeneModules.R
\name{get_modules_from_cor}
\alias{get_modules_from_cor}
\title{Get coregulation network using soft thresholding}
\usage{
get_modules_from_cor(
  cormat,
  min_module_size = 3,
  max_module_size = 20,
  resolution = 0.02,
  corthresh = 0.1,
  min_module_cor = 0.1
)
}
\arguments{
\item{cormat}{The correlation matrix}

\item{min_module_size}{An integer. Won't consider modules smaller than this.}

\item{max_module_size}{An integer. Not implemented yet.}

\item{resolution}{Argument to igraph::cluster_leiden. Lower values produce bigger clusters.}

\item{corthresh}{Only correlations about this value will go into the adjacency graph fed into leiden clustering}

\item{min_module_cor}{Only keep modules with average cor above this value.}
}
\value{
An list with two elements. \code{modules}, a list mod module memberships;
and \code{dend}, a dendrogram from hierarchical clustering of the genes
}
\description{
Derive a network graph by soft-thresholding the correlation matrix, and computing Topological Overlap Measure (TOM) values
}
