% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborhoodCalculations.R
\name{getNeighborhoodExpression}
\alias{getNeighborhoodExpression}
\title{calculate neighborhood expression}
\usage{
getNeighborhoodExpression(x, neighbors, makedense = TRUE)
}
\arguments{
\item{x}{matrix to get neighbor colMeans over}

\item{neighbors}{Sparse matrix giving neighbor relationships}

\item{makedense}{Logical, for whether to convert to dense matrix}
}
\value{
A matrix in the dimension of x holding colMeans of x over neighbors
}
\description{
calculate neighborhood expression
}
