% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPerturbation.R
\name{getPerturbations}
\alias{getPerturbations}
\title{Calculate perturbations for selected cells and genes}
\usage{
getPerturbations(
  x,
  obj,
  cells = NULL,
  genes = NULL,
  modules = NULL,
  residtype = "log2ratio",
  eps = 1
)
}
\arguments{
\item{x}{Expression matrix for complete dataset, same as was input to initializeISD. Cells in rows, genes in columns.}

\item{obj}{Output of initializeISD}

\item{cells}{Which cells to produce output for. If NULL, will return results for all cells.}

\item{genes}{Which genes to produce output for.}

\item{residtype}{Either "log2ratio" or "diff"}

\item{eps}{For log2ratio calculations, mean neighborhood expression will be thresholded below at this value}
}
\value{
Matrix of perturbation values for the selected cells vs. genes.
}
\description{
For given cells and genes, compute the matrix of perturbations from matched controls.
}
