% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePerturbation.R
\name{identifyMostPerturbedGenes}
\alias{identifyMostPerturbedGenes}
\title{Identify the most perturbed genes given output of summarizeGenePerturbation()}
\usage{
identifyMostPerturbedGenes(genescores, n = 1000)
}
\arguments{
\item{genescores}{Output of summarizeGenePerturbation()}

\item{n}{Maximum number of genes to return}
}
\value{
A vector of names of highly perturbed genes
}
\description{
Identify the most perturbed genes given output of summarizeGenePerturbation()
}
