% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPerturbation.R
\name{matchToControls}
\alias{matchToControls}
\title{Find the most similar control neighborhood(s) for each cellular neighborhood}
\usage{
matchToControls(x, tissue, iscontrol)
}
\arguments{
\item{x}{Matrix of cell x gene neighborhood expression values}

\item{tissue}{Vector of cells' tissue IDs}

\item{iscontrol}{Which cells are controls}
}
\description{
Find the most similar control neighborhood(s) for each cellular neighborhood
}
