% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minisle}
\alias{minisle}
\title{Small example CosMx data from a kidney sample}
\format{
A list with the following elements:
\itemize{
\item counts A matrix of raw counts, with cells i rows and genes in columns
\item annot A data frame holding cell metadata
\item xy A matrix of cell xy positions
\item clust A vector of cell type assignments
}
}
\usage{
minisle
}
\description{
A 3-sample, 9-FOV, 100-gene excerpt from a 1000-plex CosMx run on a kidney biopsy.
}
\keyword{datasets}
