% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pseudoRandomSample}
\alias{pseudoRandomSample}
\title{Get a random-like but deterministic subset of a vector}
\usage{
pseudoRandomSample(vec, n)
}
\arguments{
\item{vec}{Vector to subset}

\item{n}{Number of elements}
}
\value{
A subset of the vector
}
\description{
Get a random-like but deterministic subset of a vector
}
