% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGeneModules.R
\name{buildGeneModules}
\alias{buildGeneModules}
\title{Find clusters of genes perturbed in similar regions}
\usage{
buildGeneModules(
  x,
  obj,
  genes = "highlyperturbed",
  resolution = 0.02,
  corthresh = 0.1,
  min_module_cor = 0.1,
  subsetsize = 1e+05,
  residtype = "log2ratio",
  eps = 1
)
}
\arguments{
\item{x}{Expression matrix for complete dataset, same as was input to initializeISD. Cells in rows, genes in columns.}

\item{obj}{Output of initializeISD}

\item{genes}{Which genes to produce output for.
Default of "highlyperturbed" subsets to the most perturbed genes. If NULL, all genes will be used.}

\item{resolution}{Resolution parameter for leiden clustering}

\item{corthresh}{Correlations with absolute value below this will be rounded to zero to save memory}

\item{min_module_cor}{Modules must have mean correlation of at least this much to be reported}

\item{subsetsize}{Number of cells to use for these calculations}

\item{residtype}{Either "log2ratio" or "diff"}

\item{eps}{For log2ratio calculations, mean neighborhood expression will be thresholded below at this value.
Low eps will have higher sensitivity and poorer specificity for detecting perturbations in low expressers.
The default of 1 is a somewhat conservative choice.}
}
\value{
A data frame giving module name, gene name, and gene weight, for all genes included in a module.
}
\description{
Find clusters of genes perturbed in similar regions
}
