% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborhoodCalculations.R
\name{getNeighbors}
\alias{getNeighbors}
\title{function to get neighbors via all the various paths:}
\usage{
getNeighbors(xy, neighbors, tissue, k, radius, verbose)
}
\arguments{
\item{xy}{Cells' xy coordinates - a 2-column matrix aligned to the rows of mat}

\item{neighbors}{Sparse matrix of neighbor relationships. If not provided, xy will be used to calculate thi]s.}

\item{tissue}{Vector of cells' tissue IDs}

\item{k}{Number of neighbors to use to define a neighborhood}

\item{radius}{Radius used to define neighborhoods (choose this or K)}

\item{verbose}{TRUE to get more messages}
}
\value{
sparse matrix holding neighbor assignments
}
\description{
function to get neighbors via all the various paths:
}
