% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPerturbation.R
\name{initializeISD}
\alias{initializeISD}
\title{Pre-preprocessing for perturbation calculations}
\usage{
initializeISD(
  mat,
  xy,
  tissue,
  iscontrol,
  k = 100,
  radius = NULL,
  neighbor_subsample_rate = 1
)
}
\arguments{
\item{mat}{Single cell gene expression matrix. Must be non-negative. Normalized counts preferred.}

\item{xy}{Cells' xy coordinates - a 2-column matrix aligned to the rows of mat}

\item{tissue}{Vector of cells' tissue IDs}

\item{iscontrol}{Logical vector defining whether each cell belongs to a control tissue / region}

\item{k}{Number of neighbors to use to define a neighborhood}

\item{radius}{Radius used to define neighborhoods (choose this or K)}

\item{neighbors}{Sparse matrix of neighbor relationships. If not provided, xy will be used to calculate this.}

\item{residtype}{Either "log2ratio" or "diff". Default to linear-scale "diff".}

\item{makedense}{Logical, for whether to convert to dense matrix}
}
\value{
A list: \itemize{
\item neighbors Matrix giving each cell's spatial neighbors.
\item comparators Matrix giving, for each cell, its "pseudoneighbors", i.e.
the cell IDs of its comparator (control) cells' neighbors,
used to concoct a control "pseudoenvironment" for the cell.
\item tissue The tissue vector input to the function.
\item iscontrol The iscontrol vector input to the function.
}
}
\description{
Define neighbor and comparator matrices for use in calculating perturbation matrix.
Expects a whole study's data, and does fine if passed tissues with diverse disease conditions.
If your controls are relevant to different tissues (e.g. healthy liver samples as controls for liver disease
and healthy kidney controls for kidney disease), then run your data in batches
according to which control tissues are relevant.
}
