% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPerturbation.R
\name{scoreDiff}
\alias{scoreDiff}
\title{Score change from control neighborhoods}
\usage{
scoreDiff(env, controlenv, residtype, eps = 1)
}
\arguments{
\item{env}{Environment matrix of target cells}

\item{controlenv}{Environment matrix of best-matching controls}

\item{residtype}{Either "log2ratio" or "diff"}

\item{eps}{Value added to expression matrices for log2ratio differences}
}
\value{
A matrix of changes from nearest controls, in the same dimension as x
}
\description{
Calculates differences between target cell neighborhoods and their matched controls.
}
