% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePerturbation.R
\name{summarizeGenePerturbation}
\alias{summarizeGenePerturbation}
\title{Summarize genes' perturbation levels}
\usage{
summarizeGenePerturbation(
  x,
  obj,
  cells = NULL,
  residtype = "log2ratio",
  plotresults = FALSE,
  eps = 1,
  ratioeps = 0.1,
  subsetsize = 1e+05
)
}
\arguments{
\item{x}{Expression matrix for complete dataset, same as was input to initializeISD. Cells in rows, genes in columns.}

\item{obj}{Output of initializeISD}

\item{cells}{Vector of cell IDs to consider, e.g. if you want to subset to a single disease tissue.
All control cells will be included regardless. Default of NULL includes all cells.}

\item{residtype}{Either "log2ratio" or "diff"}

\item{plotresults}{Logical, for whether to make a summary plot of the results}

\item{eps}{For log2ratio calculations, this value is added to mean neighborhood expression levels.}

\item{ratioeps}{Added to means and means+2sds when calculating ratios between disease and controls.}

\item{subsetsize}{Number of cells to use for these calculations}
}
\description{
For each gene, summarize perturbation with the following statistics:
\itemize{
\item How broadly perturbed in disease: its median perturbation in non-controls / its median perturbation in controls
\item How strongly perturbed in disease: its 0.99 quantile perturbation in non-controls / its 0.999 quantile perturbation in controls
\item How variably perturbed in disease: the SD of its perturbation scores in disease
}
}
